requirements.txt
absl-py
appnope
astor
atomicwrites
attrs
backcall
bleach
certifi
chardet
Click
cloudpickle
cycler
decorator
defusedxml
dill
entrypoints
future
gast
grpcio
-e git+https://github.com/openai/gym.git@39ce693cce275e61751c8cad0ad8b6496029601f#egg=gym
h5py
html5lib
idna
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
# keras==2.11
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
more-itertools
mujoco-py
nbconvert
nbformat
notebook
numpy<2
opencv-python
pandas
pandocfilters
parso
pbr
pexpect
pickleshare
pluggy
progressbar2
prometheus-client
prompt-toolkit
protobuf
ptyprocess
py
pyglet
Pygments
PyOpenGL
pyparsing
pyrsistent
pytest
python-dateutil
python-utils
pytz
pyzmq
qtconsole
requests
scikit-learn
scipy
seaborn
Send2Trash
six
stevedore
tensorboard
tensorflow-macos==2.11
# tensorflow==2.11
termcolor
terminado
testpath
torch
tornado
tqdm
traitlets
urllib3
virtualenv
virtualenv-clone
virtualenvwrapper
wcwidth
Werkzeug
widgetsnbextension
# stable-baselines3[extra]
